Add Marionette to the mix
Not sure if I'll continue to use marionette long term but it has a lot of potential from the looks of the API in terms of cross component communication and useful utilities/defaults. Though their API is out of date and pretty terrible at actually telling you what you want to know.