First Marionette Messasging
Submit now calls through marionette to create a new account. Rejects as the feature is not yet implemented, but shell is now in place. Next step will be login in, probably from the list of known servers first.

Also added a list of known server. Will be useful in the future for detecting when we know the configuration for a server and don't have to run tests for it.