Prompt for Password
We prompt for the user's password when we try to connect to an account. Maybe at some point we'll do local encrypted passwords stored, but for now we just prompt on launch.

Like with welcome, login occurs in the background, but nothing is displayed in the UI to indicate that. Also, failure is not handled at the moment. Should reprompt eventually. Also, it will prompt multiple times I believe, when we should only prompt one at a time.