Pull out connection code from account model
Accounts don't need to be doing the IMAP connection managment themselves. Pulled out the code to a generic connection utility that is pure node. May be getting pulled into a few different places or replaced at some point, but this should keep the account model from getting bloated and make it more like a model should be.