Added Controller Object
Added a simple controller object. Unfortunately most of the pre-built ones seemed to be mostly about routing and I couldn't find a good one that was going to fit what I need. So I built a small one. Will probably grow as time goes on, but for now it will work as a simple base. Follows a lot of the same code principles that native Backbone code does.