Stop automatic disconnect
Used to disconnect after the list of messages was pulled down. This was done for so that we didn't create a bunch of open connection and hose or get locked out of an account. We will now add an 'unload' listener to disconnect (just in case). We also force disconnect during that time.