New messages are displayed in UI
The UI will now recieve and show any new messages that arrive in the inbox after we connect.

Still not quite at 0.1 version yet. Need to also handle the removal of messages, and possibly the update of message flags. Specifically read (other will likely be coming later with a bigger "flag" support push).